#3 ConversationsとMessagesのmodel作成
