+++ app/Models/Conversations.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Conversations extends Model
+{
+    use HasFactory;
+
+    protected $fillable = ['user_id', 'title'];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
+}
