+++ app/Models/Messages.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Messages extends Model
+{
+    use HasFactory;
+
+    protected $fillable = ['user_id', 'conversation_id', 'content', 'role'];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
+}
