+++ resources/views/livewire/layout/navigation.blade.php
@@ -17,7 +17,7 @@ $logout = function (Logout $logout) {
             <div class="flex">
                 <!-- Logo -->
                 <div class="shrink-0 flex items-center">
-                    <a href="{{ route('openai.index') }}" wire:navigate>
+                    <a href="{{ route('index') }}" wire:navigate>
                         <x-application-logo class="block h-9 w-auto fill-current text-gray-800" />
                     </a>
                 </div>
