+++ resources/views/livewire/pages/auth/register.blade.php
@@ -34,7 +34,7 @@ $register = function () {
 
     Auth::login($user);
 
-    $this->redirect(route('openai.index', absolute: false), navigate: true);
+    $this->redirect(route('index', absolute: false), navigate: true);
 };
 
 ?>
