+++ routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\OpenAIController;
+use App\Http\Controllers\ChatbotController;
 use Illuminate\Support\Facades\Route;
 
 // Route::view('/', 'welcome');
@@ -14,5 +14,5 @@ Route::view('profile', 'profile')
     ->name('profile');
 require __DIR__.'/auth.php';
 
-Route::get('/', [OpenAIController::class, 'index'])->middleware(['auth', 'verified'])->name('openai.index');
-Route::post('/send-message', [OpenAIController::class, 'sendMessage'])->middleware(['auth', 'verified'])->name('openai.send');
+Route::get('/', [ChatbotController::class, 'index'])->middleware(['auth', 'verified'])->name('index');
+Route::post('/chat', [ChatbotController::class, 'chat'])->middleware(['auth', 'verified'])->name('chat');
